requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
jinja2==3.1.4

# Core AI/Audio
openai-whisper==20231117
ffmpeg-python==0.2.0
numpy>=1.24.0

# Translation
transformers>=4.41.0
sentencepiece>=0.1.99
safetensors>=0.4.2
